backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
apify-client>=1.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0